+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Highcharts detection.
+- Support for gzip-encoded response body.
+
 ## [0.3.0]
 
 ### Added